fix some more warnings, and increase precision slightly for HC55516
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@3969 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/cpu/i86/instr86.c
@@ -2881,7 +2881,7 @@ static void PREFIX86(_f6pre)(void)
 				}
 				else
 				{
-					I.regs.b[AL] = result;
+					I.regs.b[AL] = (UINT8)result;
 					I.regs.b[AH] = tmp2;
 				}
 			}
